Moving all audio related code to an audio crate
