+++ amethyst_audio/src/lib.rs
@@ -1,8 +1,19 @@
 //! Loading and playing of audio files.
+extern crate amethyst_assets;
+extern crate amethyst_core;
+extern crate cgmath;
+extern crate cpal;
+extern crate rodio;
+extern crate shred;
+extern crate smallvec;
+extern crate specs;
 
 pub use self::formats::{FlacFormat, OggFormat, WavFormat};
 pub use self::sink::AudioSink;
 pub use self::source::{Source, SourceHandle};
+pub use self::components::*;
+pub use self::systems::*;
+pub use self::bundle::AudioBundle;
 
 pub mod output;
 
@@ -12,6 +23,9 @@ use std::fmt::{Display, Formatter, Result as FmtResult};
 mod formats;
 mod sink;
 mod source;
+mod components;
+mod systems;
+mod bundle;
 
 /// An error occurred while decoding the source.
 #[derive(Debug)]
