--- src/ecs/audio/mod.rs
@@ -1,9 +1,0 @@
-//! `amethyst` audio ecs module
-
-pub use self::bundle::AudioBundle;
-pub use self::components::*;
-pub use self::systems::*;
-
-pub mod components;
-pub mod systems;
-pub mod bundle;
