+++ src/ecs/mod.rs
@@ -3,8 +3,7 @@
 //! `transform` has moved to `amethyst::core::transform`
 //! `rendering` has moved to `amethyst::renderer::prelude::*` and
 //! `amethyst::renderer::bundle::RenderBundle`
-//!
+//! `audio` has moved to `amethyst::audio`
 pub use specs::*;
 
-pub mod audio;
 pub mod input;
